requirements.txt
# Core PyTorch / Hugging Face stack
torch>=2.2
transformers>=4.43
accelerate>=0.33
safetensors>=0.4
huggingface_hub>=0.24
datasets>=2.20
bitsandbytes>=0.43
numpy>=1.26
pyyaml>=6.0
requests>=2.31
tqdm>=4.66
rank-bm25>=0.2.2
ipython>=8.20
python-dotenv>=1.0
codebleu==0.6.0
sacrebleu>=2.3.1
tree_sitter>=0.21.3
tree_sitter_languages>=1.10.2